move get_user() to common_errata_tool
Move this to common_errata_tool so we can eventually use it across all
three modules that must look up user information (user, product,
release).
